requirements.txt
statsmodels
pandas
numpy
matplotlib
plotly
psycopg2-binary
python-dotenv
streamlit
bcrypt
prophet
cmdstanpy
Cython